Reorganise structure of essential analysis components and created proof of concept notebook
