+++ cell_radial_analysis/render.py
@@ -28,7 +28,7 @@ import time
 
 def auto_plot_cumulative(input_2d_hist, input_type, N, num_bins, radius, t_range, focal_cell, focal_event, subject_cell, subject_event, save_parent_dir, SI):
 
-        xlocs, xlabels, ylocs, ylabels = tools.kymo_labels(num_bins, 1, radius, t_range, SI)
+        xlocs, xlabels, ylocs, ylabels = kymo_labels(num_bins, 1, radius, t_range, SI)
 
         ## formatting cell and event names
         focal_event_name = 'apoptoses' if focal_event == 'APOPTOSIS' else 'divisions'
@@ -95,7 +95,7 @@ def auto_plot_cumulative(input_2d_hist, input_type, N, num_bins, radius, t_range
 
 def plot_cumulative(input_2d_hist, num_bins, radius, t_range, title, label, cb_label, save_path, SI):
 
-        xlocs, xlabels, ylocs, ylabels = tools.kymo_labels(num_bins, 1, radius, t_range, SI)
+        xlocs, xlabels, ylocs, ylabels = kymo_labels(num_bins, 1, radius, t_range, SI)
 
         if SI == True:
             time_unit = '(Hours)'
@@ -132,7 +132,7 @@ def plot_N_cells(input_2d_hist, subject_cells, target_cell, focal_time, radius, 
 
     num_bins = len(input_2d_hist)
 
-    xlocs, xlabels, ylocs, ylabels = tools.kymo_labels(num_bins, 2, radius, t_range, SI = False)
+    xlocs, xlabels, ylocs, ylabels = kymo_labels(num_bins, 2, radius, t_range, SI = False)
 
     #expt_label = 'expt:' + expt_ID + '\n 90:10 WT:Scr\n'
     #plt.text(num_bins+1,num_bins+4,expt_label)
@@ -140,18 +140,18 @@ def plot_N_cells(input_2d_hist, subject_cells, target_cell, focal_time, radius, 
     plt.yticks(ylocs, ylabels)
     plt.xlabel("Time since t = " + str(focal_time)+ ' (frames)')
     plt.ylabel("Distance from focal event (pixels)")
-    plt.title('Kymograph for '+target_cell.fate.name+' ' +cell_type+' ID:'+str(cell_ID)+" at t= "+ str(focal_time))
+    plt.title('Kymograph for '+target_cell.fate.name.lower()+' ' +cell_type+' ID:'+str(cell_ID)+" at t= "+ str(focal_time))
 
     final_hist = np.flipud(input_2d_hist)  ## flip for desired graph orientation
     plt.imshow(final_hist)
 
     if min([cell.ID for cell in subject_cells]) > 0:
-        cb_label = 'Probability of wild-type mitoses'
+        cb_label = 'Number of wild-type cells'
     if min([cell.ID for cell in subject_cells]) < 0:
-        cb_label = 'Probability of Scribble mitoses? AMEND THIS LABEL'
+        cb_label = 'Number of Scribble cells? AMEND THIS LABEL'
 
-    if event == 'APOPTOSIS':
-        raise Exception('Apoptosis event counter not configured yet')
+    # if event == 'APOPTOSIS':
+    #     raise Exception('Apoptosis event counter not configured yet')
 
     plt.colorbar(label = cb_label)
 
@@ -167,7 +167,7 @@ def plot_N_events(input_2d_hist, event, subject_cells, target_cell, focal_time, 
 
     num_bins = len(input_2d_hist)
 
-    xlocs, xlabels, ylocs, ylabels = tools.kymo_labels(num_bins, 2, radius, t_range, SI = False)
+    xlocs, xlabels, ylocs, ylabels = kymo_labels(num_bins, 2, radius, t_range, SI = False)
 
     #expt_label = 'expt:' + expt_ID + '\n 90:10 WT:Scr\n'
     #plt.text(num_bins+1,num_bins+4,expt_label)
@@ -201,7 +201,7 @@ def plot_P_events(input_2d_hist, event, subject_cells, target_cell, focal_time, 
 
     num_bins = len(input_2d_hist)
 
-    xlocs, xlabels, ylocs, ylabels = tools.kymo_labels(num_bins, 2, radius, t_range)
+    xlocs, xlabels, ylocs, ylabels = kymo_labels(num_bins, 2, radius, t_range)
 
 
     #expt_label = 'expt:' + expt_ID + '\n 90:10 WT:Scr\n'
@@ -225,3 +225,28 @@ def plot_P_events(input_2d_hist, event, subject_cells, target_cell, focal_time, 
     plt.colorbar(label = cb_label)
 
     return plt.imshow(final_hist)
+
+
+### labels in SI
+def kymo_labels(num_bins, label_freq, radius, t_range, SI):
+    label_freq =1
+    radial_bin = radius / num_bins
+    temporal_bin = t_range / num_bins
+
+    if SI == True:
+        time_scale_factor = 4/60 ## each frame is 4 minutes
+        distance_scale_factor = 1/3 ## each pixel is 0.3recur micrometers
+    else:
+        time_scale_factor, distance_scale_factor = 1,1
+
+    ### generate labels for axis micrometers/hours
+    xlocs = range(0, num_bins,label_freq) ## step of 2 to avoid crowding
+    xlabels = []
+    for m in range(int(-num_bins/2), int(num_bins/2),label_freq):
+        xlabels.append(str(int(((temporal_bin)*m)*time_scale_factor)) + "," + str(int(((temporal_bin)*m+temporal_bin)*time_scale_factor)))
+    ylocs = range(0, num_bins, label_freq) ## step of 2 to avoid crowding
+    ylabels = []
+    for m in range(num_bins, 0, -label_freq):
+        ylabels.append(str(int(((radial_bin)*m)*distance_scale_factor)) + "," + str(int(((radial_bin)*(m-1)*distance_scale_factor))))
+
+    return xlocs, xlabels, ylocs, ylabels
